Add support for bold and italic text + window 5 theme (#30)
* Update theme.lua

* Update ui.lua

* added themes based on window 5 style

Tried to make themes look as much like the base game's window 5 style as I could. Both translucent (like the game) and solid versions included.

* testing new prompts and positioning

Fiddling with the text and prompt offsets, and trying out different prompt icons. The ones from the game I was using are too pixelated, and can't be animated properly (yet) so they don't look right.

* changed text size and style

Made message text smaller and changed stroke color/width to be more consistent with base game. Made npc text slightly smaller to match.

* changed prompt and added alternates

Lots of experimentation with prompt styles. Settling on plain white arrow for now, but including alternate prompts in case I change my mind or a user wants something different.

* replaced name background

This new background uses the whole canvas so it should be easier to style using the theme xml. This also means less editing in gimp if someone wants to make minor changes.

* adjusted message stroke

Text was a bit too blurry so I dropped the stroke width a bit and decreased the alpha.

* Update theme.xml

Copied the message stroke settings to the system block.

* changed balloon background and name background

The old balloon style was too rigid; the gap in the balloon's border looked nice but makes it impossible to adjust npc name positioning without having to update the background image as well. This one doesn't look as nice, but it's much easier to work with. Adjusted npc name background to compensate so the name text is still easy to read.

* prompt adjustment

Changed the size and position of the prompt icon slightly to match the "more text" icon for equipment details.

* copied translucent changes to solid theme

The solid theme is just the regular theme without alpha, so copied all the new art and the style sheet over for consistency. Also created a new background with a solid border, for the same reasons as the default version.

* added readmes to themes

Added readmes to explain my reasoning for the design, and how the themes were made. I can also add change notes to the readmes later, if and when I make changes to the themes.

---------

Co-authored-by: civilityreigns <100637768+civilityreigns@users.noreply.github.com>